service/requirements.txt
Flask==1.0.2
xmltodict
pytz
simple-salesforce
iso8601
python-dateutil
